Hide model.config from JSON, fix config update
Each `ModelDefinition` keeps a reference to its `ModelConfig`
in `config` property. This commit makes that property non-enumerable,
so that it is not sent to the server and not stored in `model.json`.

The second part of this commit fixes an bug where ModelConfig changes
were not persisted when the model name was changed at the same time.