client/test: infrastructure for MySQL datasource
Allow integration tests to use a MySQL database for operations that
require a datasource (a connector) with automigrate and autodiscovery
support.

Create a gulp task to setup a MySQL database and credentials to be used
by the tests.

Add a smoke test running autoupdate to verify the setup works as
expected.

Fix issues discovered along the way:
 - `given.modelInstance` was ignoring `modelConfig` parameter
 - `ModelService.migrateModelConfig` had inconsistent return type,
   either a promise or a $resource object.