Improve metrics background stability
Improve the stability of the metrics by cleaning up the objects used to hold charting data when the page is put into the background.

Unselect the current process, and stop the update timer to prevent any
repeating processes from running while the tab is in the background.
When the tab regains focus, reselect the previous active process and
reload the graphs.