test: get protractor tests working again
- make login/logout buttons more identifiable for testing
- bump timeouts for CI
- don't load arc until after fs cleanup

The mesh proxy starts polling as soon as it is required, which is when
server/server.js is required. The test server needs to
require('../../server/server') AFTER it has finished its synchronous FS
cleanup, otherwise we use up most or all of the 5 second connection
timeout between arc and the PM before Arc or the PM have even been
started.