build-deploy: use exit status for success/fail
Code used to be assuming that any stderr lines indicated build failure,
and reporting an error to the user. Despite the error, the build would
continue in the background and often succeed.

Correct strategy is to use sl-build exit status. The build "flakiness"
is easily reproduceable by running `DEBUG=strong-build* slc arc`,
because debug output goes to stderr, it would cause all builds to be
considered to have failed.

Other fixes:

- handle spawn error events
- use debug library so that component is debuggable
- some deletion of dead code and fixing of comment typos
- sorting of require statements